# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: goapp API
  version: 0.1.0
  description: >
    Hand-authored OpenAPI spec — the single source of truth for the HTTP contract.
    The backend serves it at /openapi.yaml (docs UI at /docs); the frontend
    generates its TypeScript types from it. A contract test
    (test/integration/openapi_test.go) validates real responses against this spec
    to catch drift.
servers:
  - url: /
...
paths:
...
  /api/v1/users/:
    get:
      tags: [users]
      summary: List users (authenticated directory)
      parameters:
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserListEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/articles/:
    get:
      tags: [articles]
      summary: List articles
      security: []
      parameters:
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ArticleStatus" }
        - name: author_id
          in: query
          schema: { type: integer, format: int64 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: A page of articles.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleListEnvelope" }
    post:
      tags: [articles]
      summary: Create an article
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateArticleRequest" }
      responses:
        "201":
          description: The created article.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    Skip:
      name: skip
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...
  schemas:
    Permissions:
      type: object
      additionalProperties: { type: boolean }
      description: Per-record action permissions for the current actor.
    User:
      type: object
      required: [id, email, role, first_name, last_name, is_active, is_verified, created_at]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
        first_name: { type: string }
        last_name: { type: string }
        is_active: { type: boolean }
        is_verified: { type: boolean }
        created_at: { type: string, format: date-time }
        permissions: { $ref: "#/components/schemas/Permissions" }
    ArticleStatus:
      type: string
      enum: [draft, published, archived]
    Article:
      type: object
      required: [id, author_id, title, content, status, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        author_id: { type: integer, format: int64 }
        title: { type: string }
        content: { type: string }
        status: { $ref: "#/components/schemas/ArticleStatus" }
        published_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        permissions: { $ref: "#/components/schemas/Permissions" }
...
    CreateArticleRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 200 }
        content: { type: string }
...
    PaginationMeta:
      type: object
      required: [skip, limit, filtered_count, total_count]
      properties:
        skip: { type: integer }
        limit: { type: integer }
        filtered_count: { type: integer, format: int64 }
        total_count: { type: integer, format: int64 }
...
    ArticleEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Article" }
    UserListEnvelope:
      allOf:
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/User" }
        - $ref: "#/components/schemas/PaginationMeta"
    ArticleListEnvelope:
      allOf:
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Article" }
        - $ref: "#/components/schemas/PaginationMeta"
...
    ErrorBody:
      type: object
      required: [message, message_key]
      properties:
        message: { type: string }
        message_key: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorBody" }
  responses:
...
    Unauthorized:
      description: Authentication required or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ValidationError:
      description: Validation failed (per-field messages in error.fields).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
security:
  - bearerAuth: []